requirements-all.txt
absl-py==2.1.0
aiobotocore==2.12.1
aiofiles==23.2.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
botocore==1.34.51
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
flake8==7.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
iniconfig==2.0.0
jmespath==1.0.1
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
pip==24.0
pip-chill==1.0.3
pluggy==1.4.0
protobuf==4.25.3
pycodestyle==2.11.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.1.1
pytest-asyncio==0.23.5.post1
pytest-cov==4.1.0
pytest_httpserver==1.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.28.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
